requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==2.1.0
sqlalchemy==2.0.32
websocket-client==1.8.0
statsmodels==0.14.2
plotly==5.24.0
